+++ includes/class-plugin.php
@@ -14,6 +14,7 @@ class PLT_Plugin
     private PLT_Club_Map $club_map;
     private PLT_Standings_Service $standings_service;
     private PLT_TheSportsDB_Client $thesportsdb_client;
+    private PLT_WPLL_Client $wpll_client;
     private PLT_Shortcode $shortcode;
     private PLT_Next_Match_Shortcode $next_match_shortcode;
 
@@ -33,9 +34,13 @@ class PLT_Plugin
         $this->api_client = new PLT_Api_Client();
         $this->club_map = new PLT_Club_Map();
         $this->thesportsdb_client = new PLT_TheSportsDB_Client();
+        $this->wpll_client = new PLT_WPLL_Client();
         $this->standings_service = new PLT_Standings_Service(
             new PLT_Football_Data_Provider($this->api_client),
-            new PLT_TheSportsDB_Provider(),
+            [
+                new PLT_WPLL_Standings_Provider($this->wpll_client),
+                new PLT_TheSportsDB_Provider(),
+            ],
             $this->club_map
         );
         $this->shortcode = new PLT_Shortcode($this->settings, $this->standings_service);
@@ -44,7 +49,8 @@ class PLT_Plugin
             $this->next_match_settings,
             $this->api_client,
             $this->standings_service,
-            $this->thesportsdb_client
+            $this->thesportsdb_client,
+            $this->wpll_client
         );
 
         add_action('plugins_loaded', [$this, 'boot']);